move child storage functions into its own module
