+++ test/testers/rust-tester/src/pdre_api/utils/mod.rs
@@ -1,10 +1,12 @@
 mod crypto;
 mod network;
 mod storage;
+mod child_storage;
 
 pub use crypto::CryptoApi;
 pub use network::NetworkApi;
 pub use storage::StorageApi;
+pub use child_storage::ChildStorageApi;
 
 // Convenience function:
 // Gets the Wasm blob which was generated by the `build.rs` script
